Add Add Products Files
Added Add-Products.html and Add-Products.js, started working on the design.

Added the controller to the app.js file and to the index.html file (as a Script).

Moved the Add-Product.html file out of the auth folder in order to be able to see it without any issues, will be moved back to auth after it is completed.