feat(email-backend): add email endpoint on api  (#52)
* feat(email backend): add email endpoint

handles POST requests, validates input, sends emails and store log in database

* refactor(email endpoint): add routing and email templates

closes #29

* feat(email-backend): add email log table and fix db migration

EmailLog is a table that keeps track of every email sent. It logs whether an email succeeded or
failed, and provides a reason to why it failed. The email provider used is SendGrid. All email
configurations is present in .env .

* refactor(email-backend): move email configuration to .env